Fixed the indexing issue in the leaderboard screen
